requirements.txt
python-redmine==1.5.1
termcolor==1.1.0
